test_requirements.txt
django-nose
django_libs
coverage
django-coverage
ipdb
flake8
fabric
factory_boy
mock
